nutrition_database/requirements.txt
flask==3.0.3
gunicorn==22.0.0
tensorflow==2.15.0
Pillow==10.3.0
numpy==1.26.4
werkzeug==3.0.3
torch
transformers
